wallets auth

+++ src/wallets/wallets.repository.ts
@@ -8,26 +8,31 @@ import { UpdateWalletDto } from './dto/update-wallet.dto';
 export class WalletsRepository {
   constructor(private prisma: PrismaService) {}
 
-  async findAll(): Promise<Wallet[]> {
-    return this.prisma.wallet.findMany();
+  async findAll(userId: string): Promise<Wallet[]> {
+    return this.prisma.wallet.findMany({ where: { user_id: userId } });
   }
 
-  async findOneById(id: string): Promise<Wallet | null> {
-    return this.prisma.wallet.findUnique({ where: { id } });
+  async findOneById(id: string, userId: string): Promise<Wallet | null> {
+    return this.prisma.wallet.findUnique({ where: { id, user_id: userId } });
   }
 
-  async create(createWalletDto: CreateWalletDto): Promise<Wallet | null> {
-    return this.prisma.wallet.create({ data: createWalletDto });
+  async create(
+    createWalletDto: CreateWalletDto,
+    userId: string,
+  ): Promise<Wallet | null> {
+    return this.prisma.wallet.create({
+      data: { ...createWalletDto, user_id: userId },
+    });
   }
 
-  async update(id: string, updateWalletDto: UpdateWalletDto) {
+  async update(id: string, updateWalletDto: UpdateWalletDto, userId: string) {
     return this.prisma.wallet.update({
-      where: { id },
+      where: { id, user_id: userId },
       data: updateWalletDto,
     });
   }
 
-  async delete(id: string): Promise<Wallet | null> {
-    return this.prisma.wallet.delete({ where: { id } });
+  async delete(id: string, userId: string): Promise<Wallet | null> {
+    return this.prisma.wallet.delete({ where: { id, user_id: userId } });
   }
 }
